Fix For Loading Files with Will To Fight Editor
-An old version of returnGamePath() was included in the code for the
WillToFight editor. Since the original creation of this class bugs in
this function have been fixed and instead of having a copy in each class
one universal returnGamePath() is now in the Sabtool_Settings class.

To fix this problem the existing returnGamePath() in the WillToFight
class has been removed and all calls to it have been set to call the
version found in SabTool_Settings.